requirements.txt
requests==2.23.0
wheel==0.34.2
python-dateutil==2.8.1
werkzeug==0.15.5
flask_bcrypt==0.7.1
typing-extensions==3.7.4.2
more_itertools==8.3.0
psycopg2==2.8.5
cryptography==2.9.2
pyTelegramBotAPI==3.7.1
airtable-python-wrapper==0.13.0
apache-airflow==1.10.10